agent/requirements.txt
spoon-ai-sdk
openai
python-dotenv
fastapi
uvicorn[standard]
pydantic
